Deadline times to long

+++ Java/src/model/ProjectManagementModel.java
@@ -2,6 +2,7 @@ package model;
 
 public interface ProjectManagementModel
 {
+
     void addProject(String name, String description);
     Project[] getProjectsByName(String projectName);
     Project[] getAllProjects();
@@ -59,7 +60,7 @@ public interface ProjectManagementModel
     Task[] getTasksByName(Requirement requirement, String name);
     void deleteTask(Requirement requirement, Task task);
     void changeTask(Task task, String name, int estimatedTime, String description,
-                           int deadlineTime, TeamMember responsibleTeamMember);
+                           long deadlineTime, TeamMember responsibleTeamMember);
     Task[] getTasksByStatus(Requirement requirement, String status);
     Task[] getTasksDaysBeforeDeadline(Requirement requirement, int days);
     int getTaskID(Task task);
@@ -76,7 +77,7 @@ public interface ProjectManagementModel
     void setName(Task task, String name);
     void setDescription(Task task, String description);
     void setEstimatedTime(Task task, int estimatedTime);
-    void setDeadlineTIme(Task task, int time);
+    void setDeadlineTIme(Task task, long time);
     void setResponsibleTeamMember(Task task, TeamMember teamMember);
     void setStatus(Task task, String status);
     void setTimeWorked(Task task, TeamMember teamMember, int time);
